Level Management System and Tester Abstract Script
Created a level management system. You can call first level, next level, or restart level, and it resets everything there. It also has checks to make sure nothing breaks logic. Internally, it stores a current level, which contains the gameObject prefab to be spawned, the name of the level, the music to be played during that level, and the level to be played upon completion of that level (can also be a cutscene).
Also, set up a better testing framework where all test classes extend from a "Tester" class that automatically enabled or disables debug messages.